# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Assist Api",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/private/notes/shortlist": {
            "get": {
                "tags": [
                    "Notes private"
                ],
                "description": "Get notes as list",
                "operationId": "22e9664618ff1c3ef050f2eada9b3068",
                "parameters": [
                    {
                        "parameter": "Authorization",
                        "name": "Authorization",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "default": "Bearer <authorization_token>"
                        }
                    },
                    {
                        "parameter": "Content-type",
                        "name": "Content-type",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "default": "application/json"
                        }
                    },
                    {
                        "parameter": "Accept",
                        "name": "Accept",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "default": "application/json"
                        }
                    },
                    {
                        "name": "search",
                        "in": "query",
                        "required": false
                    }
                ],
                "responses": {
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "msg": {
                                            "type": "string",
                                            "example": "Unauthorized"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/NotePrivateShortResource"
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "msg": {
                                            "type": "string",
                                            "example": "Server Error"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/private/notes/{id}": {
            "get": {
                "tags": [
                    "Notes private"
                ],
                "description": "Get note by id",
                "operationId": "0bbbbebf43b845bfa7b8149bd3c3bb67",
                "parameters": [
                    {
                        "parameter": "Authorization",
                        "name": "Authorization",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "default": "Bearer <authorization_token>"
                        }
                    },
                    {
                        "parameter": "Content-type",
                        "name": "Content-type",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "default": "application/json"
                        }
                    },
                    {
                        "parameter": "Accept",
                        "name": "Accept",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "default": "application/json"
                        }
                    },
                    {
                        "name": "{id}",
                        "in": "path",
                        "description": "ID model",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "default": 1
                        }
                    }
                ],
                "responses": {
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "msg": {
                                            "type": "string",
                                            "example": "Unauthorized"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/NotePrivateFullResource"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "msg": {
                                            "type": "string",
                                            "example": "Server Error"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Notes private"
                ],
                "description": "Update note",
                "operationId": "6b6792ff2b3e08426a0bbe35024db183",
                "parameters": [
                    {
                        "parameter": "Authorization",
                        "name": "Authorization",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "default": "Bearer <authorization_token>"
                        }
                    },
                    {
                        "parameter": "Content-type",
                        "name": "Content-type",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "default": "application/json"
                        }
                    },
                    {
                        "parameter": "Accept",
                        "name": "Accept",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "default": "application/json"
                        }
                    },
                    {
                        "name": "{id}",
                        "in": "path",
                        "description": "ID model",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "default": 1
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/NoteRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "msg": {
                                            "type": "string",
                                            "example": "Unauthorized"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/NotePrivateFullResource"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "The requested data was not found.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "msg": {
                                            "type": "string",
                                            "example": "Not Found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Invalid data",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "msg": {
                                            "type": "string",
                                            "example": "Some error message"
                                        },
                                        "errors": {
                                            "properties": {
                                                "field": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string"
                                                    },
                                                    "example": [
                                                        "validation error by this field"
                                                    ],
                                                    "nullable": false
                                                }
                                            },
                                            "type": "object",
                                            "nullable": false
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "msg": {
                                            "type": "string",
                                            "example": "Server Error"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Notes private"
                ],
                "description": "Delete note",
                "operationId": "d338de420443770c481c2b2ce83f747d",
                "parameters": [
                    {
                        "parameter": "Authorization",
                        "name": "Authorization",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "default": "Bearer <authorization_token>"
                        }
                    },
                    {
                        "parameter": "Content-type",
                        "name": "Content-type",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "default": "application/json"
                        }
                    },
                    {
                        "parameter": "Accept",
                        "name": "Accept",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "default": "application/json"
                        }
                    },
                    {
                        "name": "{id}",
                        "in": "path",
                        "description": "ID model",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "default": 1
                        }
                    }
                ],
                "responses": {
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "msg": {
                                            "type": "string",
                                            "example": "Unauthorized"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "204": {
                        "description": "Success"
                    },
                    "404": {
                        "description": "The requested data was not found.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "msg": {
                                            "type": "string",
                                            "example": "Not Found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "msg": {
                                            "type": "string",
                                            "example": "Server Error"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "NoteRequest": {
                "required": [
                    "title"
                ],
                "properties": {
                    "title": {
                        "type": "string",
                        "example": "some title",
                        "nullable": false
                    },
                    "text": {
                        "example": "some text",
                        "nullable": true,
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ]
                    },
                    "tags": {
                        "type": "array",
                        "items": {
                            "type": "integer"
                        },
                        "example": [
                            1,
                            2
                        ],
                        "nullable": false
                    },
                    "links": {
                        "description": "Adding links related to a notes",
                        "type": "array",
                        "items": {
                            "properties": {
                                "title": {
                                    "type": "string",
                                    "example": "Original text",
                                    "nullable": false
                                },
                                "link": {
                                    "type": "string",
                                    "example": "https://example.com",
                                    "nullable": false
                                }
                            },
                            "type": "object"
                        },
                        "nullable": false
                    }
                },
                "type": "object",
                "additionalProperties": false
            },
...
            "NotePrivateFullResource": {
                "required": [
                    "id",
                    "status",
                    "title",
                    "slug",
                    "text",
                    "weight",
                    "created_at",
                    "tags",
                    "links",
                    "meta"
                ],
                "properties": {
                    "id": {
                        "description": "Model id",
                        "type": "integer",
                        "example": 22,
                        "nullable": false
                    },
                    "status": {
                        "type": "string",
                        "example": "draft",
                        "nullable": false
                    },
                    "title": {
                        "type": "string",
                        "example": "database",
                        "nullable": false
                    },
                    "slug": {
                        "type": "string",
                        "example": "database",
                        "nullable": false
                    },
                    "text": {
                        "type": "string",
                        "example": "some text",
                        "nullable": false
                    },
                    "weight": {
                        "type": "integer",
                        "example": 8,
                        "nullable": false
                    },
                    "created_at": {
                        "type": "string",
                        "example": "2025-02-02 11:49",
                        "nullable": false
                    },
                    "tags": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/TagResource"
                        },
                        "nullable": false
                    },
                    "links": {
                        "type": "array",
                        "items": {
                            "properties": {
                                "title": {
                                    "type": "string",
                                    "example": "Original text",
                                    "nullable": false
                                },
                                "link": {
                                    "type": "string",
                                    "example": "https://example.com",
                                    "nullable": false
                                }
                            },
                            "type": "object"
                        },
                        "nullable": false
                    },
                    "meta": {
                        "description": "Данные по модели, которые помогают обрабатывать ее",
                        "properties": {
                            "statuses": {
                                "description": "Статусы на котороые можно переключуть данную модель",
                                "type": "array",
                                "items": {
                                    "properties": {
                                        "value": {
                                            "type": "string",
                                            "example": "draft",
                                            "nullable": false
                                        },
                                        "label": {
                                            "type": "string",
                                            "example": "Draft",
                                            "nullable": false
                                        }
                                    },
                                    "type": "object"
                                },
                                "nullable": false
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object",
                "additionalProperties": false
            },
            "NotePrivateShortResource": {
                "required": [
                    "id",
                    "title"
                ],
                "properties": {
                    "id": {
                        "description": "Model id",
                        "type": "integer",
                        "example": 22,
                        "nullable": false
                    },
                    "title": {
                        "type": "string",
                        "example": "database",
                        "nullable": false
                    }
                },
                "type": "object",
                "additionalProperties": false
            },
...
            "TagResource": {
                "required": [
                    "id",
                    "name",
                    "slug",
                    "color",
                    "weight"
                ],
                "properties": {
                    "id": {
                        "description": "Model id",
                        "type": "integer",
                        "example": 22,
                        "nullable": false
                    },
                    "name": {
                        "type": "string",
                        "example": "database",
                        "nullable": false
                    },
                    "slug": {
                        "type": "string",
                        "example": "database",
                        "nullable": false
                    },
                    "color": {
                        "type": "string",
                        "example": "#d98b84",
                        "nullable": false
                    },
                    "weight": {
                        "type": "integer",
                        "example": 4,
                        "nullable": false
                    }
                },
                "type": "object",
                "additionalProperties": false
            },
...
        }
    },
...
}